+++ src/commands/aircraft/adirs.ts
@@ -1,8 +1,8 @@
 import { MessageCommandDefinition } from '../../lib/command';
 import { makeEmbed, makeLines } from '../../lib/embed';
-import { CommandCategory } from '../../constants';
+import { CommandCategory, imageBaseUrl } from '../../constants';
 
-const ADIRS_IMAGE_URL = `${process.env.IMAGE_BASE_URL}a32nx/adirs.png`;
+const ADIRS_IMAGE_URL = `${imageBaseUrl}/a32nx/adirs.png`;
 
 const adirsEmbed = makeEmbed({
     title: 'FlyByWire A32NX | ADIRS align',