+++ src/commands/general/notams.ts
@@ -1,4 +1,4 @@
-import { CommandDefinition } from '../../lib/command';
+import { CommandDefinition, replyWithMsg } from '../../lib/command';
 import { CommandCategory } from '../../constants';
 
 const WEB_BASE_URL = 'https://flybywiresim.com';
@@ -7,5 +7,5 @@ export const notams: CommandDefinition = {
     name: ['notams', 'notam', 'news'],
     description: 'Links to the FlyByWire Simulations NOTAMS.',
     category: CommandCategory.GENERAL,
-    executor: (msg) => msg.channel.send(`${WEB_BASE_URL}/notams`),
+    executor: (msg) => replyWithMsg(msg, `${WEB_BASE_URL}/notams`),
 };