+++ src/commands/memes/noHello.ts
@@ -1,3 +1,4 @@
+import { Message } from 'discord.js';
 import { CommandDefinition } from '../../lib/command';
 import { CommandCategory } from '../../constants';
 
@@ -7,8 +8,13 @@ export const noHello: CommandDefinition = {
     name: ['nohello', 'hello'],
     description: 'No!',
     category: CommandCategory.MEMES,
-    executor: (msg) => {
-        msg.channel.send(NO_HELLO_URL);
-        return msg.delete();
+    executor: async (msg: Message) => {
+        const sentMsg = await msg
+            .fetchReference()
+            .then((res) => res.reply(NO_HELLO_URL))
+            .catch(async () => msg.channel.send(NO_HELLO_URL));
+
+        msg.delete();
+        return sentMsg;
     },
 };